sonarqube: remove unnecessary casts, simplify conditionals, unused fields, redundant null checks, NOSONAR exclusions (S1905/S1126/S1450/S4201/S3012/S3878, 35 fixes)

+++ router/java/src/net/i2p/router/tasks/CoalesceStatsEvent.java
@@ -142,9 +142,9 @@ public class CoalesceStatsEvent implements SimpleTimer.TimedEvent {
         sm.addRateData("router.memoryUsed", used);
         if (_maxMemory - used < LOW_MEMORY_THRESHOLD) {Router.clearCaches();}
 
-        sm.addRateData("router.cpuLoad", (long) SystemVersion.getCPULoad());
-        sm.addRateData("tunnel.tunnelBuildSuccessAvg", (long) SystemVersion.getTunnelBuildSuccess());
-        sm.addRateData("router.activeThreads", (int) SystemVersion.getActiveThreads());
+        sm.addRateData("router.cpuLoad", SystemVersion.getCPULoad());
+        sm.addRateData("tunnel.tunnelBuildSuccessAvg", SystemVersion.getTunnelBuildSuccess());
+        sm.addRateData("router.activeThreads", SystemVersion.getActiveThreads());
 
         _ctx.tunnelDispatcher().updateParticipatingStats(Router.COALESCE_TIME);
